pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "empujon-llm"
version = "1.0.0"
description = "Unified LLM Multiplexer for the Empujón Educativo ecosystem"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
openai = ["openai>=1.0"]
gemini = ["google-genai>=1.0"]
bedrock = ["boto3>=1.28"]
all = ["openai>=1.0", "google-genai>=1.0", "boto3>=1.28"]

[tool.setuptools.packages.find]
include = ["empujon_llm*"]
